production start

+++ database/migrations/2021_11_25_070043_create_process_number.php
@@ -18,7 +18,7 @@ class CreateProcessNumber extends Migration
             $table->integer('json_id')
             ->nullable(false);
             $table->text('key');
-            $table->float('value');
+            $table->double('value');
 
             $table->timestamps();
         });
